ScientificSystem/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.8
astroid==1.5.3
autopep8==1.3.2
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
crypto==1.4.1
cymysql==0.9.1
Django==3.0.7
django-cors-headers==3.5.0
dwebsocket==0.5.12
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Cors==2.1.0
Flask-HTTPAuth==2.7.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.5
isort==4.2.15
itchat==1.3.9
itsdangerous==0.24
Jinja2==2.9.6
kombu==3.0.37
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.4.6
Naked==0.1.31
paho-mqtt==1.2
Pillow==7.2.0
pyasn1==0.3.1
pycodestyle==2.3.1
pylint==1.7.2
PyMuPDF==1.17.7
PyMySQL==0.9.3
PyPDF2==1.26.0
pypng==0.0.18
PyQRCode==1.2.1
pytz==2017.2
PyYAML==3.12
qrcode==5.3
redis==2.10.5
reportlab==3.5.49
requests==2.18.3
rsa==3.4.2
shellescape==3.4.1
simplejson==3.17.0
six==1.10.0
SQLAlchemy==1.1.13
sqlparse==0.3.1
urllib3==1.22
users==1.0.dev0
vine==1.1.4
Werkzeug==0.12.2
wrapt==1.10.10
WTForms==2.1
XlsxWriter==1.3.3
xlwt==1.3.0
